requirements.txt
streamlit~=1.36.0
yfinance~=0.2.40
pandas~=2.2.2
matplotlib~=3.9.1
google-generativeai
tessa
streamlit-antd-components
plotly
newsapi-python
nltk
tensorflow
keras
scikit-learn